requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
openai==1.51.2
python-dateutil==2.9.0.post0
